Implement Assembler::compileR

+++ wasmc++/include/wasm/Registers.h
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <array>
+#include <cstdint>
 #include <map>
 
 namespace Wasmc {
-	extern std::map<const std::string *, int> registerMap;
+	extern std::map<const std::string *, uint64_t> registerMap;
 	extern std::array<const std::string *, 128> registerArray;
 }
