+++ wasmc++/src/wasm/Registers.cpp
@@ -2,7 +2,7 @@
 #include "wasm/Registers.h"
 
 namespace Wasmc {
-	std::map<const std::string *, int> registerMap {
+	std::map<const std::string *, uint64_t> registerMap {
 		{StringSet::intern("$0"),    0},
 		{StringSet::intern("$g"),    1},
 		{StringSet::intern("$sp"),   2},
